Revert ":sparkles: Refactor a few things, major porting progress"
This reverts commit 334935957645188f1c9e230aeb51fcb1c24425a8.

+++ src/src-code/system/dolphin/types.h
@@ -3,12 +3,12 @@
 
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+typedef unsigned long u32;
 typedef unsigned long long u64;
-typedef signed char s8;
-typedef signed short s16;
-typedef signed int s32;
-typedef signed long long s64;
+typedef char s8;
+typedef short s16;
+typedef long s32;
+typedef long long s64;
 typedef float f32;
 typedef double f64;
 typedef volatile s64 vs64;
@@ -22,7 +22,7 @@ typedef volatile u8 vu8;
 typedef volatile f64 vf64;
 typedef volatile f32 vf32;
 
-typedef unsigned long size_t;
+typedef unsigned int size_t;
 typedef int BOOL;
 
 #define NULL 0
